+++ src/main/java/bowling/model/KnockedDownPins.java
@@ -1,5 +1,6 @@
 package bowling.model;
 
+import bowling.model.knockdownstrategy.KnockDownNumber;
 import java.util.Objects;
 
 public class KnockedDownPins {
