+++ src/main/java/bowling/model/PlayerName.java
@@ -1,6 +1,5 @@
 package bowling.model;
 
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class PlayerName {
