+++ src/main/java/bowling/model/framestatus/FrameStatus.java
@@ -1,5 +1,6 @@
-package bowling.model;
+package bowling.model.framestatus;
 
+import bowling.model.KnockedDownPins;
 import java.util.List;
 
 public interface FrameStatus {
