+++ src/main/java/bowling/model/knockdownstrategy/FirstKnockDownStrategy.java
@@ -1,4 +1,6 @@
-package bowling.model;
+package bowling.model.knockdownstrategy;
+
+import bowling.model.KnockedDownPins;
 
 public class FirstKnockDownStrategy implements KnockDownStrategy {
